+++ benchmark/exampleCkksPerformanceDefault.js
@@ -72,8 +72,8 @@ function create() {
     process.stdout.write(
       `Done [${Math.round((timeEnd - timeStart) * 1000)} microseconds]\r\n`
     )
-    const secretKey = keyGenerator.getSecretKey()
-    const publicKey = keyGenerator.getPublicKey()
+    const secretKey = keyGenerator.secretKey()
+    const publicKey = keyGenerator.publicKey()
 
     const relinKeys = seal.RelinKeys()
     const galoisKeys = seal.GaloisKeys()
@@ -81,14 +81,14 @@ function create() {
     if (context.usingKeyswitching) {
       process.stdout.write('Generating relinearization keys: ')
       timeStart = performance.now()
-      relinKeys.move(keyGenerator.genRelinKeys())
+      relinKeys.move(keyGenerator.relinKeysLocal())
       timeEnd = performance.now()
       process.stdout.write(
         `Done [${Math.round((timeEnd - timeStart) * 1000)} microseconds]\r\n`
       )
       process.stdout.write('Generating Galois keys: ')
       timeStart = performance.now()
-      galoisKeys.move(keyGenerator.genGaloisKeys())
+      galoisKeys.move(keyGenerator.galoisKeysLocal())
       timeEnd = performance.now()
       process.stdout.write(
         `Done [${Math.round((timeEnd - timeStart) * 1000)} microseconds]\r\n`