+++ scripts/seal-clean.sh
@@ -1,9 +1,19 @@
 #!/usr/bin/env bash
 
-rm -rf ./src/bin/js/* ./src/bin/wasm/*  \
-&& cd ./submodules/SEAL/native \
-&& rm -rf ./bin/* \
-&& rm -rf ./lib/*  \
-&& rm -rf ./src/CMakeFiles \
-&& rm -rf ./src/CMakeCache.txt \
-&& cd ../../../
+rm -rf ./src/bin/js/* ./src/bin/wasm/*
+cd ./submodules/SEAL/
+rm -rf ./bin/*
+rm -rf ./lib/*
+rm -rf ./CMakeFiles
+rm -rf ./CMakeCache.txt
+rm -rf ./native/src/CMakeFiles
+rm -rf ./native/src/CMakeCache.txt
+rm -rf ./thirdparty/zlib/CMakeFiles
+rm -rf ./thirdparty/zlib/CMakeCache.txt
+rm -rf ./thirdparty/zlib/src/CMakeFiles
+rm -rf ./thirdparty/zlib/src/CMakeCache.txt
+rm -rf ./thirdparty/msgsl/CMakeFiles
+rm -rf ./thirdparty/msgsl/CMakeCache.txt
+rm -rf ./thirdparty/msgsl/src/CMakeFiles
+rm -rf ./thirdparty/msgsl/src/CMakeCache.txt
+cd ../../